# chore(remap): Improve the `Display` performance of `Kind`