--- a/lib/vrl/compiler/Cargo.toml
+++ b/lib/vrl/compiler/Cargo.toml
@@ -28,3 +28,7 @@ criterion = "0.3"
 [[bench]]
 name = "path"
 harness = false
+
+[[bench]]
+name = "kind"
+harness = false