--- a/lib/vrl/compiler/src/value/kind.rs
+++ b/lib/vrl/compiler/src/value/kind.rs
@@ -68,6 +68,14 @@ impl Kind {
         self == self.scalar()
     }
 
+    /// Returns a quoted variant of `as_str`
+    ///
+    /// This function is a close duplicate of `as_str`, returning the same
+    /// underlying str but with quotes. We avoid the obvious `format!("{}",
+    /// self.as_str())` here as that incurs an allocation cost and the `Display`
+    /// of `Kind` is sometimes in the hot path.
+    ///
+    /// See https://github.com/timberio/vector/pull/6878 for details.
     pub(crate) fn quoted(self) -> &'static str {
         match self {
             Kind::Bytes => "\"string\"",