@blt 2021-03-24
@jszwedko good point about the comment. I'll add one. 

Regarding the path, couldn't tell you. `coz` just suggests lines that could profit overall throughput but there's no information about what paths were hit in the program that led to that conclusion.