@jszwedko 2021-03-24
> @jszwedko good point about the comment. I'll add one.
> 
> Regarding the path, couldn't tell you. `coz` just suggests lines that could profit overall throughput but there's no information about what paths were hit in the program that led to that conclusion.

👍 I just wasn't sure if you had a hunch or might be able to tell using a "traditional profiler". The reason I'm wondering is perhaps whatever it is might be able to avoid calling `quoted()` at all. I might have guessed we'd only see that called for debug or error output.

Interestingly, the observed speed up here of roughly 3% does not match coz's prediction of a 50% total speed-up based on the roughly 10% improvement to the line speed. It's first data point is at 25% line speed up though so my guess is the "line fitting" isn't reflective of reality for 10%. Or there is something else going on.