@blt 2021-03-24
> +1 I just wasn't sure if you had a hunch or might be able to tell using a "traditional profiler". The reason I'm wondering is perhaps whatever it is might be able to avoid calling quoted() at all. I might have guessed we'd only see that called for debug or error output.

No, I wish. It'd be fantastically useful if we could get that information, I do agree. I think the challenge with call graph analysis like that is it's just very, very expensive for large programs. 

> Interestingly, the observed speed up here of roughly 3% does not match coz's prediction of a 50% total speed-up based on the roughly 10% improvement to the line speed. It's first data point is at 25% line speed up though so my guess is the "line fitting" isn't reflective of reality for 10%. Or there is something else going on.

Indeed. I was disappointed by the all-up results, but them's the breaks. It's entirely possible that the path didn't trigger as much in subsequent runs or by optimizing this one hotspot we traded for another. I'd like to do a tear down of `coz` here soon. I wonder if we didn't have enough points available for this path to really draw a good conclusion.