@JeanMertz 2021-03-25
This is definitely interesting and curious.

@blt it would be good to have a chat sometime this week, so that I get a better understanding of what is actually being benchmarked in https://github.com/timberio/vector/pull/6868.

Some early notes:

- `Kind`'s string representation is/should (as @jszwedko pointed out) only used for debug purposes and generating compiler (or runtime, if opted into by the program) error messages.
- Compiler performance has never been a major priority for the project, given that it isn't a hot path (only runs once at boot). Instead, our time was spent on runtime performance (and everything else to get v1 out the door).
- Given this, I can see four things going on here:
   1. `kind.quoted()` (or any function this uses, such as `kind.to_string()`, etc) somehow leaked into a VRL runtime hot-path.
   2. We're somehow compiling the program more than once after Vector booted, which would be a severe problem.
   3. Running the benchmarks results in runtime type errors that trigger the error-path which heavily uses this function.
   4. The benchmarks are focussing on compiler performance, which might show localized performance issues that don't have a significant impact on actual usage.

If (1) is the case, then we should track that down. I would be shocked if (2) was true, but it's still worth making sure this isn't the case. I don't believe (3) or (4) are the case, but it's again worth keeping in mind.

It's also worth mentioning (since I know you've been dabbling with it lately) that the VRL CLI is a very unoptimized "hobby" project that I built in a couple of hours to make it easy to debug VRL programs, and was deemed useful enough to be shipped in 0.12 for others to use. It got _some_ love by others before we shipped it, but I would still consider it a "non-production" utility, especially when it comes to performance.

Re: getting better insights into what happens when. Another item that I never got around to due to time constraints, but that we still should do at some point, is adding more tracing/profiling to VRL itself. We've already got profiling possibilities in Vector through the `tracing` crate (among others), but nothing like that is currently available in VRL, while it would help us figure out if/when functions like these are called erroneously.