@JeanMertz 2021-03-25
I did a brief check on (1), but couldn't find any uses other than in error messages (but this might not have caught all possible uses):

```
% rg --no-heading --with-filename "kind\.quoted" lib/vrl/
lib/vrl/compiler/src/value/kind.rs:143:                    kinds.push(Kind::$kind.quoted())

% rg --no-heading --with-filename "kind\.to_string" lib/vrl/
lib/vrl/compiler/src/expression/function_call.rs:517:                        kind.to_string()

% rg --no-heading --with-filename "format!\(.*kind.*" lib/vrl/
lib/vrl/compiler/src/expression/predicate.rs:147:                Label::context(format!("instead it resolves to {}", kind), self.span),
lib/vrl/compiler/src/expression/not.rs:99:                    format!("this expression resolves to {}", kind),
lib/vrl/compiler/src/expression/function_call.rs:519:                        format!(r#"the exact type {}"#, kind)
lib/vrl/compiler/src/expression/function_call.rs:521:                        format!("one of {}", kind)
lib/vrl/compiler/src/expression/function_call.rs:527:                        format!("this expression resolves to {}", kind_str(got)),
lib/vrl/stdlib/src/float.rs:52:            v => Err(format!(r#"expected "float", got {}"#, v.kind()).into()),
lib/vrl/stdlib/src/timestamp.rs:52:            v => Err(format!(r#"expected "timestamp", got {}"#, v.kind()).into()),
lib/vrl/stdlib/src/to_int.rs:115:            v => Err(format!(r#"unable to coerce {} into "integer""#, v.kind()).into()),
lib/vrl/stdlib/src/to_float.rs:116:            v => Err(format!(r#"unable to coerce {} into "float""#, v.kind()).into()),
lib/vrl/stdlib/src/array.rs:52:            v => Err(format!(r#"expected "array", got {}"#, v.kind()).into()),
lib/vrl/stdlib/src/to_timestamp.rs:139:            v => return Err(format!(r#"unable to coerce {} into "timestamp""#, v.kind()).into()),
lib/vrl/stdlib/src/string.rs:52:            v => Err(format!(r#"expected "string", got {}"#, v.kind()).into()),
lib/vrl/stdlib/src/object.rs:52:            v => Err(format!(r#"expected "object", got {}"#, v.kind()).into()),
lib/vrl/stdlib/src/to_string.rs:104:            v => return Err(format!(r#"unable to coerce {} into "string""#, v.kind()).into()),
lib/vrl/stdlib/src/integer.rs:52:            v => Err(format!(r#"expected "integer", got {}"#, v.kind()).into()),
lib/vrl/stdlib/src/boolean.rs:52:            v => Err(format!(r#"expected "boolean", got {}"#, v.kind()).into()),
lib/vrl/stdlib/src/to_bool.rs:161:            v => Err(format!(r#"unable to coerce {} into "boolean""#, v.kind()).into()),
```